refactor(ui): unify book cover card components into a single BookCoverCard with variants
